requirements.txt
# Server
Flask==3.0.0
gunicorn==21.2.0
Werkzeug==3.0.1
gevent==23.9.1

# Production dependencies
python-dotenv==1.0.0
pymongo==4.6.2
flask-pymongo==2.3.0
python-jose==3.3.0
bcrypt==4.1.2 